requirements.txt
langchain==0.2.14
langchain-groq===0.1.9
langchain-community==0.2.12
streamlit==1.35.0
python-dotenv==1.0.0
unstructured==0.14.6
pandas==2.0.2
selenium==4.21.0
chromadb==0.5.0